Resharper moves
