+++ src/Humans.Web/Controllers/AdminDuplicateAccountsController.cs
@@ -4,7 +4,6 @@ using Humans.Web.Authorization;
 using Humans.Web.Models;
 using Humans.Application;
 using Humans.Application.Interfaces.Teams;
-using Humans.Application.Interfaces.Profiles;
 using Humans.Application.Interfaces.Users;
 
 namespace Humans.Web.Controllers;
