+++ tests/Humans.Application.Tests/Services/Gdpr/GdprExportDependencyInjectionTests.cs
@@ -8,7 +8,7 @@ using ApplicationDecisionService = Humans.Application.Services.Governance.Applic
 using BudgetService = Humans.Application.Services.Budget.BudgetService;
 using ExpenseReportService = Humans.Application.Services.Expenses.ExpenseReportService;
 using CampaignService = Humans.Application.Services.Campaigns.CampaignService;
-using ProfilesAccountMergeService = Humans.Application.Services.Profiles.AccountMergeService;
+using ProfilesAccountMergeService = Humans.Application.Services.Users.AccountMergeService;
 using UsersUserService = Humans.Application.Services.Users.UserService;
 using AuditLogService = Humans.Application.Services.AuditLog.AuditLogService;
 using CampService = Humans.Application.Services.Camps.CampService;
