+++ tests/Humans.Integration.Tests/AccountMerge/AcceptAsyncFoldTests.cs
@@ -1,6 +1,6 @@
 using AwesomeAssertions;
-using Humans.Application.Interfaces.Profiles;
 using Humans.Application.Interfaces.Repositories;
+using Humans.Application.Interfaces.Users;
 using Humans.Domain.Constants;
 using Humans.Domain.Entities;
 using Humans.Domain.Enums;
