+++ tests/Humans.Integration.Tests/AccountMerge/ChainFollowReadTests.cs
@@ -3,7 +3,7 @@ using Humans.Application.Interfaces.AuditLog;
 using Humans.Application.Interfaces.Budget;
 using Humans.Application.Interfaces.Consent;
 using Humans.Application.Interfaces.Gdpr;
-using Humans.Application.Interfaces.Profiles;
+using Humans.Application.Interfaces.Users;
 using Humans.Domain.Constants;
 using Humans.Domain.Entities;
 using Humans.Domain.Enums;
